+++ EasyKeyKit/EasyKeyKit.swift
@@ -1,5 +1,5 @@
 import Foundation
 
 public enum EasyKeyKit {
-    public static let version = "0.0.13"
+    public static let version = "0.0.14"
 }